# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AniDB API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/catalogs": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Get All Catalog",
                "operationId": "bf7d3c3af4d71538d00b0df849f3d908",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Catalog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Add a Catalog",
                "operationId": "8d851d66dad56921006fe15cca087444",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "2020"
                    },
                    {
                        "name": "season",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Winter",
                                "Spring",
                                "Summer",
                                "Fall"
                            ]
                        },
                        "example": "Winter"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/catalogs/{catalog_id}": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Get Partials in Catalog",
                "operationId": "4b0774dc654f92416d46f3d7a10d20b4",
                "parameters": [
                    {
                        "name": "catalog_id",
                        "in": "path",
                        "description": "Catalog ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "e9597119-8452-4f2b-96d8-f2b1b1d2f158"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Partial"
                                            }
                                        },
                                        "stats": {
                                            "$ref": "#/components/schemas/Catalog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Edit a Catalog",
                "operationId": "0eab08efecfd0eb04f1f6805da8af369",
                "parameters": [
                    {
                        "name": "catalog_id",
                        "in": "path",
                        "description": "Catalog ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "e9597119-8452-4f2b-96d8-f2b1b1d2f158"
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "2020"
                    },
                    {
                        "name": "season",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Winter",
                                "Spring",
                                "Summer",
                                "Fall"
                            ]
                        },
                        "example": "Winter"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Delete a Catalog",
                "operationId": "cfdc75fe92bd8e71c989ed0696b8a53f",
                "parameters": [
                    {
                        "name": "catalog_id",
                        "in": "path",
                        "description": "Catalog ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "example": "e9597119-8452-4f2b-96d8-f2b1b1d2f158"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Catalog": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "year": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "season": {
                        "type": "string",
                        "enum": [
                            "Winter",
                            "Spring",
                            "Summer",
                            "Fall"
                        ]
                    }
                },
                "type": "object",
                "example": {
                    "id": "e9597119-8452-4f2b-96d8-f2b1b1d2f158",
                    "year": 2020,
                    "season": "Winter"
                }
            },
...
            "Partial": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "High",
                            "Normal",
                            "Low"
                        ]
                    }
                },
                "type": "object",
                "example": {
                    "id": "e9597119-8452-4f2b-96d8-f2b1b1d2f158",
                    "title": "Sample Title",
                    "priority": "High"
                }
            },
...
        },
        "responses": {
            "Success": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "format": "int32"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "status": 200,
                            "message": "Success"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "format": "int32"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": {
                            "status": 401,
                            "message": "Unauthorized"
                        }
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}